Add verify tokens for home_timeline and post_tweet
previously if you run the server and write in browser URl "http://localhost:8000/home_timeline" OR "http://localhost:8000/post_tweet" then you can directly visit that page whether you authorize or not . To avoid this add to check access token. If the access token is present then you can go to that page. but if not then it will redirect to login page.